requirements.txt
selenium
bs4
spacy
unidecode
python-dateutil
python-dotenv
psycopg
rapidfuzz
tqdm
pytest
git+ssh://git@github.com/jtai-dev/record_matcher.git#egg=record_matcher